+++ NEWS.md
@@ -3,6 +3,7 @@
 * `leveneTests()` gains formula and lm methods
 * `bartlettTests()` gains formula and lm methods
 * `colDevs()` gains a `group.var` argument; examples extended
+* Added `uniStats()` for an anova version of `glance.mlm()`
 
 ## Version 1.7.2 (2024-11-25)